Piece placement logic
Added logic for the piece placement. You can place all of the pieces on both black and white. Need to look into not placing a piece on a file with just the king a pawn. Also, coving up the other side of the board when black is placing their pieces.